+++ packages/pyroscope-flamegraph/src/FlameGraph/FlameGraphComponent/color.ts
@@ -93,6 +93,8 @@ function spyToRegex(spyName: string) {
       return /^(?<packageName>(.*\/)*)(?<filename>.*\.rb+)(?<line_info>.*)$/;
     case 'nodespy':
       return /^(\.\/node_modules\/)?(?<packageName>[^/]*)(?<filename>.*\.?(jsx?|tsx?)?):(?<functionName>.*):(?<line_info>.*)$/;
+    case 'tracing':
+      return /^(?<packageName>.+?):.*$/;
     case 'javaspy':
       // TODO: we might want to add ? after groups
       return /^(?<packageName>.+\/)(?<filename>.+\.)(?<functionName>.+)$/;
@@ -125,13 +127,13 @@ export function colorBasedOnPackageName(
   palette: FlamegraphPalette,
   name: string
 ) {
-  const hash = murmurhash3_32_gc(name);
+  const hash = murmurhash3_32_gc(name, 0);
   const colorIndex = hash % palette.colors.length;
   const baseClr = palette.colors[colorIndex];
   if (!baseClr) {
     console.warn('Could not calculate color. Defaulting to the first one');
     // We assert to Color since the first position is always available
-    return palette.colors[0] as Color;
+    return palette.colors[0];
   }
 
   return baseClr;
@@ -154,7 +156,7 @@ export function NewDiffColor(
       goodColor.rgb().toString(),
       neutralColor.rgb().toString(),
       badColor.rgb().toString(),
-    ] as any);
+    ] as ShamefulAny);
 
   return (n: number) => {
     // convert to our Color object