+++ webapp/javascript/pages/AdhocDiff.tsx
@@ -99,6 +99,7 @@ function AdhocDiff() {
           )}
           {!isProfileLoading && (
             <FlamegraphRenderer
+              showCredit={false}
               flamebearer={flamebearer}
               ExportData={
                 <ExportData