+++ webapp/javascript/redux/reducers/user.ts
@@ -2,7 +2,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { type User } from '@webapp/models/users';
 import { connect, useSelector } from 'react-redux';
-import { PAGES } from '@webapp/pages/constants';
 import {
   loadCurrentUser as loadCurrentUserAPI,
   changeMyPassword as changeMyPasswordAPI,
@@ -55,7 +54,7 @@ const userSlice = createSlice({
   reducers: {},
   extraReducers: (builder) => {
     builder.addCase(loadCurrentUser.fulfilled, (state, action) => {
-      return { type: 'loaded', data: action.payload as User };
+      return { type: 'loaded', data: action.payload };
     });
     builder.addCase(loadCurrentUser.pending, (state) => {
       return { type: 'loading', data: state.data };