+++ packages/pyroscope-panel-plugin/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.19](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.18...@pyroscope/panel-plugin@1.3.19) (2022-07-11)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.18](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.17...@pyroscope/panel-plugin@1.3.18) (2022-07-11)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.3.17](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.16...@pyroscope/panel-plugin@1.3.17) (2022-07-11)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
 ## [1.3.16](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.3.15...@pyroscope/panel-plugin@1.3.16) (2022-07-07)
 
 **Note:** Version bump only for package @pyroscope/panel-plugin